restructure project, add preferences, hook into uv menu
* add .gitignore
* split files
* add preferences
* pref option to add BRM UVTools to UV Menu, either as individual entries or as a submenuA
* add header text for operators
* minor cleanup and renaming
